Re #307 Fixed GCC 4.4 more warnings after r4487
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@4493 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: deda396f3873d59b430125fec8eb122e4b5ba41f
Former-commit-id: 04195fb242a228634dad0b0daba84de27817dc9e

+++ src/modules/graphic/ssggraph/grsound.cpp
@@ -92,13 +92,13 @@ void grInitSound(tSituation* s, int ncars)
 		rpm_scale = GfParmGetNum(handle, "Sound", "rpm scale", NULL, 1.0);
 		sprintf(filename, "cars/%s/%.*s",
 			car->_carName,
-			sizeof(filename) - strlen(car->_carName) - strlen("cars//"),
+			(int)(sizeof(filename) - strlen(car->_carName) - strlen("cars//")),
 			param);
 		file = fopen(filename, "r");
 		if (!file)
 		{
  			sprintf(filename, "data/sound/%.*s",
-				sizeof(filename) - strlen(car->_carName) - strlen("data/sound/"),
+				(int)(sizeof(filename) - strlen(car->_carName) - strlen("data/sound/")),
 				param);
 		}
 		else
